fuzzing: faster marshalling and comparison-phase due to less IO (#96)
Currently, all evms output is parsed, into `StructLog` items. Those are then lightly cleansed, and then marshalled into json again, and written to output files. 

In the next phase, those output files are then compared. This comparison is done line by line, and does not parse json, just compares the raw bytes. 

## Custom marshaller 

This PR changes the output / comparison phase. First of all, the 'standard' json marshaller is discarded, and instead a custom-made marshaller is used. This has three benefits

- It is much faster
- The custom marshaller removes the need for removing content from the log objects in a pre-step, since we can just omit emitting things we're not concerned about, 
- It only prints the top six stack items, making the output more condensed and less intensive to write and read. 


Example comparison-data before this PR: 
```json
{"pc":0,"op":0,"gas":"0x0","gasCost":"0x0","memSize":0,"stack":["0x0","0x1","0x2","0x3","0x4","0x5","0x6","0x7","0x8","0x9"],"depth":0,"refund":0,"opName":"STOP"}

```
Example comparison data with this PR: 
```json
 {"depth":0,"pc":0,"gas":0,"op":0,"opName":"STOP","stack":["0x4","0x5","0x6","0x7","0x8","0x9"]}
```
`162` vs `95` bytes on disk. This PR initially replaced `json` with an even more compact form, but reverted that, in order to make it interoperable with tools like `traceview`. 


## Avoid diskwrites

However, a more interesting optimization was added, where for the most part we don't even touch the disk. We use a `dualWriter`, and feeds the output above into it. 

The `dualWriter` does two things; writes into a `hash`, and writes to a buffered writer. The buffered writer has a capacity of 5Mb, if that gets full, it goes down to disk. After we're written everything, we compare the hashes for all evms. If the hashes match, we discard all buffered content, and don't read the files. Only if the hashes mismatch, do we actually flush the (last) 5MB of buffered data down to disk, and do the line-by-line comparison. 

For very large traces, there will be some disk-writes performed. 
For consensus-flaws, there will be both disk-writes and disk-readbacks. But most executions will never touch disk!