Faster execution of fuzzing-pipeline (#121)
This PR implements https://github.com/holiman/goevmlab/issues/120. This creates a huge speed-boost in the case where many clients are used. 

Previously, the general flow of execution was

- Generate a test
- Spin up `n` clients to execute it, 
- Wait for all to complete
- Compare the outputs
- Repeat

That is, every test is executed by `n` clients, and the slowest one sets the pace. With this PR, we instead do 

- Generate a test
- Wait until we have two clients ready, task them to execute it, 
- Whenever a client is done, handle it's output, put into ready-queue
- Repeat

That is, every test is executed by `2` clients, the first two clients that are available. Whenever a client becomes available, it is tasked with a new test shortly. 

So instead of the slowest client setting the pace, each client can execute in their own pace, and they get a corresponding workload. Faster clients get more test coverage.