docker: separate nimbus-eth1 out, remove rust after installation of eels (#209)
This PR splits away the nimbus-eth1 evm client into a separate image, and also cleans up the rust stuff in the main container.

The split makes it so we don't have to rebuild nimbus-eth1 each time the 'main' is rebuilt (it takes a huge amount of time), and the latter change makes the image a lot++ smaller.

With this PR, the size has gone from 2.65GB to 1.66GB