Ability to debug raw output (`--rawdebug`) (#211)
This adds a flag for a new feature, which can be used when consensus-flaws appear that are very hard to reprodue. That is, consensus flaws that are temporary, and do not appear twice in a row.

Normally, goevmlab funnel the vm outputs into hashers, and once the n (usually two) clients are done executing, compares the two hashes. If they mismatch, then a consensus-flaw is present, and the tests is re-ran on all clients, and the outputs compared.

And normally, this works fine. However, it can sometimes happen that the error only appears the first time, and when we run them all together, the error is no longer present.

This PR adds a flag --rawdebug. It makes the hasher actually hold on to the raw vm output. If we enter the 'consensus-error' mode, it spits out the n outputs from the clients under test into files.